requirements.txt
discord.py==1.3.4

requests==2.24.0

bs4==0.0.1

lxml==4.5.2

PyNaCl==1.4.0
